Add missing `<cstdint>` include to utf8_validity_fuzzer.cc
This will allow us to move utf8_range over to our main CI infrastructure in
preparation for merging it into the protobuf repo.

I ended up needing to include a few other fixes as well. To get our Mac tests
(which use Python 3.12) passing, I had to upgrade our `rules_python` version.
This in turn required upgrading to Bazel 6, which broke some things in
`rules_fuzzing` which also had to be fixed. I'm working on upstreaming the
`rules_fuzzing` fixes so that we don't have to carry a patch for long.

PiperOrigin-RevId: 579968688